Fix misuse of {} initialization
Narrowing is supposed to be forbidden in {} initialization, but Chromium
doesn't appear to care upstream.

+++ chromium/device/fido/public_key_credential_params.cc
@@ -39,7 +39,8 @@ PublicKeyCredentialParams::CreateFromCBORValue(const cbor::Value& cbor_value) {
     }
 
     credential_params.push_back(PublicKeyCredentialParams::CredentialInfo{
-        CredentialType::kPublicKey, algorithm_type_it->second.GetInteger()});
+        CredentialType::kPublicKey,
+        static_cast<int>(algorithm_type_it->second.GetInteger())});
   }
 
   return PublicKeyCredentialParams(std::move(credential_params));
