+++ chromium/printing/print_settings_conversion.cc
@@ -90,7 +90,7 @@ PageRanges GetPageRangesFromJobSettings(const base::Value& job_settings) {
 
       // Page numbers are 1-based in the dictionary.
       // Page numbers are 0-based for the printing context.
-      page_ranges.push_back(PageRange{from.value() - 1, to.value() - 1});
+      page_ranges.push_back(PageRange{uint32_t(from.value() - 1), uint32_t(to.value() - 1)});
     }
   }
   return page_ranges;
